- adding tests;

+++ src/main/java/net/jfaenomoto/mutant/model/StatusResponse.java
@@ -38,7 +38,7 @@ public class StatusResponse {
         if (this.humans + this.mutants == 0) {
             return 0.0;
         }
-        return this.mutants / (this.humans + this.mutants);
+        return (double) this.mutants / (this.humans + this.mutants);
     }
 
 }
